+++ DesignToolbox/DesignToolbox/DesignToolbox.swift
@@ -20,6 +20,10 @@ struct DesignToolbox: App {
     @StateObject private var themeProvider = ThemeProvider()
     @AppStorage("colorSchemeMode") private var mode: String = ColorSchemeMode.auto.rawValue
 
+    #if os(macOS)
+    @StateObject private var windowManager = WindowManager()
+    #endif
+
     private var colorScheme: ColorScheme? {
         if mode == ColorSchemeMode.light.rawValue {
             return .light
@@ -36,6 +40,9 @@ struct DesignToolbox: App {
                 MainView().environmentObject(themeProvider)
             }
             .preferredColorScheme(colorScheme)
+            #if os(macOS)
+                .environmentObject(windowManager)
+            #endif
         }
     }
 }