+++ DesignToolbox/DesignToolbox/Pages/About/AccessibilityStatement/AccessibilityStatementPage.swift
@@ -11,6 +11,11 @@
 // Software description: A SwiftUI components library with code examples for Orange Unified Design System
 //
 
+// NOTE: Seen as unused but are used
+// This is a false positive in Periphy
+// See https://github.com/peripheryapp/periphery/issues/908
+
+#if os(iOS)
 import DeclarationAccessibility
 import OUDSFoundations
 import SwiftUI
@@ -39,3 +44,4 @@ struct AccessibilityStatementPage: View {
         }
     }
 }
+#endif