+++ DesignToolbox/DesignToolbox/Pages/Components/Controls/Radio/RadioElements.swift
@@ -44,8 +44,12 @@ private struct RadioIllustration: View {
     var body: some View {
         VStack(spacing: theme.spaces.fixedNone) {
             // 🥜: dumb label, not vocalized, preventing warnings because of empty labels
+            #if !os(visionOS)
             OUDSRadio(isOn: .constant(true), accessibilityLabel: "🥜")
             OUDSRadio(isOn: .constant(false), accessibilityLabel: "🥜")
+            #else
+            OUDSRadio(isOn: .constant(false), accessibilityLabel: "🥜")
+            #endif
         }
     }
 }