+++ DesignToolbox/DesignToolbox/Pages/Utils/DesignToolbox+EnrichedTexts.swift
@@ -14,6 +14,10 @@
 import OUDS
 import SwiftUI
 
+// NOTE: Several items below are seen as unused but are used
+// This is a false positive in Periphy
+// See https://github.com/peripheryapp/periphery/issues/908
+
 // MARK: - Copyable Text View Modifier
 
 /// A  `ViewModifier` defining a context menu allowing user to copy in clipboard some content.
@@ -30,7 +34,7 @@ struct CopyableTextViewModifier: ViewModifier {
         content
             .contextMenu {
                 Button(action: {
-                    UIPasteboard.general.string = copyable
+                    OSUtilities.copy(content: copyable)
                 }, label: {
                     Text("app_common_copy" <- copyable)
                     Image(systemName: "doc.on.doc").accessibilityHidden(true)
@@ -91,7 +95,7 @@ struct OpenableText: View {
                 .underline()
                 .oudsForegroundStyle(theme.link.colorContentEnabled)
                 .onTapGesture {
-                    UIApplication.shared.open(URL(string: type.destination(for: anchor).first!)!)
+                    OSUtilities.open(url: type.destination(for: anchor).first!)
                 }
                 .accessibilityAddTraits([.isLink])
         } else {
@@ -102,7 +106,7 @@ struct OpenableText: View {
                         .underline()
                         .oudsForegroundStyle(theme.link.colorContentEnabled)
                         .onTapGesture {
-                            UIApplication.shared.open(url)
+                            OSUtilities.open(url: url)
                         }
                         .accessibilityAddTraits([.isLink])
                 } else {