+++ DesignToolbox/DesignToolbox/Utils/View+extensions.swift
@@ -1,0 +1,33 @@
+//
+// Software Name: OUDS iOS
+// SPDX-FileCopyrightText: Copyright (c) Orange SA
+// SPDX-License-Identifier: MIT
+//
+// This software is distributed under the MIT license,
+// the text of which is available at https://opensource.org/license/MIT/
+// or see the "LICENSE" file for more details.
+//
+// Authors: See CONTRIBUTORS.txt
+// Software description: A SwiftUI components library with code examples for Orange Unified Design System
+//
+
+import SwiftUI
+
+extension View {
+
+    func textInputDisableAutocapitalization() -> some View {
+        #if os(iOS)
+        return textInputAutocapitalization(.never)
+        #else
+        return self
+        #endif
+    }
+
+    func numberPadKeyboard() -> some View {
+        #if os(iOS)
+        return keyboardType(.numberPad)
+        #else
+        return self
+        #endif
+    }
+}